[Silver III] Title: 주유소, Time: 380 ms, Memory: 43708 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -6,8 +6,8 @@ public class Main {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
 		int N = Integer.parseInt(br.readLine());
-		int[] length = new int[N - 1];
-		int[] price = new int[N];
+		long[] length = new long[N - 1];
+		long[] price = new long[N];
 
 		StringTokenizer st = new StringTokenizer(br.readLine(), " ");
 
@@ -23,14 +23,14 @@ public class Main {
 
 		int pre = 0;
 		int next = 0;
-		int result = length[pre] * price[pre];
+		long result = length[pre] * price[pre];
 
 		for (int i = 1; i < N - 1; i++) {
 			next = i;
 			if (price[pre] > price[next]) {
 				pre = i;
 			}
-
+			
 			result += length[i] * price[pre];
 		}
